backend/requirements.txt
Django==5.2.1
djangorestframework
django-cors-headers
django-filter
psycopg2-binary
python-dotenv
requests
gunicorn
